requirements.txt
pyTelegramBotAPI
tweepy
instaloader
requests
urllib3
bilibili-api-python
yt-dlp
beautifulsoup4
python-dotenv
